rename some functions
