+++ chain/src/chain.rs
@@ -979,7 +979,7 @@ impl BlockChain {
             vm_metrics,
         )?;
         let (state_root2, included_txn_info_hashes2) = if !transactions2.is_empty() {
-            let (state_root, hashes) = starcoin_vm2_chain::execute_vm2_txns_and_save(
+            let (state_root, hashes) = starcoin_vm2_chain::execute_txns_and_save(
                 to_hash_value2(block_id),
                 block.header.number(),
                 storage2,
