+++ vm2/executor/src/lib.rs
@@ -1,6 +1,11 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod executor2;
+pub mod executor;
 
-pub mod block_executor2;
+pub mod block_executor;
+
+pub use block_executor::{block_execute, execute_genesis_transaction};
+pub use executor::{
+    do_execute_block_transactions, execute_readonly_function, validate_transaction,
+};
